requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
Babel>=1.3
boto>=2.32.1
boto3>=1.2.2
decorator>=3.4.0
fixtures>=1.3.1
iso8601>=0.1.9
Jinja2>=2.6 # BSD License (3 clause)
jsonschema!=2.5.0,<3.0.0,>=2.0.0
netaddr!=0.7.16,>=0.7.12
oslo.config>=2.6.0 # Apache-2.0
oslo.db>=3.0.0 # Apache-2.0
oslo.i18n>=1.5.0 # Apache-2.0
oslo.log>=1.8.0 # Apache-2.0
oslo.serialization>=1.10.0 # Apache-2.0
oslo.utils!=2.6.0,>=2.4.0 # Apache-2.0
paramiko>=1.13.0
pbr>=1.6
PrettyTable<0.8,>=0.7
PyYAML>=3.1.0
python-designateclient>=1.5.0
python-glanceclient>=0.18.0
python-keystoneclient!=1.8.0,>=1.6.0
python-novaclient>=2.29.0
python-neutronclient>=2.6.0
python-cinderclient>=1.3.1
python-manilaclient>=1.3.0
python-heatclient>=0.6.0
python-ceilometerclient>=1.5.0
python-ironicclient>=0.8.0
python-saharaclient>=0.10.0
python-troveclient>=1.2.0
python-zaqarclient>=0.1.1
python-swiftclient>=2.2.0
python-subunit>=0.0.18
requests!=2.8.0,>=2.5.2
SQLAlchemy<1.1.0,>=0.9.9
sphinx!=1.2.0,!=1.3b1,<1.3,>=1.1.2
six>=1.9.0

# Python 2.6 related packages(see rally.common.costilius for more details)
ordereddict
simplejson>=2.2.0
